feat: new go file structure
