+++ api/internal/bittensor/crypto.go
@@ -1,0 +1,48 @@
+package bittensor
+
+import (
+	"crypto/sha256"
+	"encoding/hex"
+	"log"
+
+	"github.com/ChainSafe/go-schnorrkel"
+)
+
+func signMessage(message []byte, public string, private string) string {
+	// Signs a message via schnorrkel pub and private keys
+
+	var pubk [32]byte
+	data, err := hex.DecodeString(public)
+	if err != nil {
+		// TODO
+		log.Fatalf("Failed to decode public key: %s", err)
+	}
+	copy(pubk[:], data)
+
+	var prik [32]byte
+	data, err = hex.DecodeString(private)
+	if err != nil {
+		log.Fatalf("Failed to decode private key: %s", err)
+	}
+	copy(prik[:], data)
+
+	priv := schnorrkel.SecretKey{}
+	_ = priv.Decode(prik)
+	pub := schnorrkel.PublicKey{}
+	_ = pub.Decode(pubk)
+
+	signingCtx := []byte("substrate")
+	signingTranscript := schnorrkel.NewSigningContext(signingCtx, message)
+	sig, _ := priv.Sign(signingTranscript)
+	sigEncode := sig.Encode()
+	out := hex.EncodeToString(sigEncode[:])
+
+	return "0x" + out
+}
+
+func sha256Hash(str []byte) string {
+	h := sha256.New()
+	h.Write(str)
+	sum := h.Sum(nil)
+	return hex.EncodeToString(sum)
+}
