+++ api/server.go
@@ -1,75 +1,31 @@
 package main
 
 import (
-	"database/sql"
-	"fmt"
-	"os"
-	"strconv"
-	"strings"
-	"time"
+	"api/internal/config"
+	"api/internal/routes"
+	"api/internal/shared"
 
 	"github.com/aidarkhanov/nanoid"
-	_ "github.com/go-sql-driver/mysql"
-	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
-	"github.com/redis/go-redis/v9"
 	"go.uber.org/zap"
 )
 
-var (
-	HOTKEY           string
-	PUBLIC_KEY       string
-	PRIVATE_KEY      string
-	INSTANCE_UUID    string
-	DEBUG            bool
-	FALLBACK_API_KEY string
-
-	REDIS_CLIENT *redis.Client
-	SQL_CLIENT   *sql.DB
-)
-
-var (
-	Reset  = "\033[0m"
-	Red    = "\033[31m"
-	Green  = "\033[32m"
-	Yellow = "\033[33m"
-	Blue   = "\033[34m"
-	Purple = "\033[35m"
-	Cyan   = "\033[36m"
-	Gray   = "\033[37m"
-	White  = "\033[97m"
-)
-
-type Context struct {
-	echo.Context
-	log   *zap.SugaredLogger
-	reqid string
-}
-
 func main() {
-	HOTKEY = safeEnv("HOTKEY")
-	PUBLIC_KEY = safeEnv("PUBLIC_KEY")
-	PRIVATE_KEY = safeEnv("PRIVATE_KEY")
-	FALLBACK_API_KEY = safeEnv("FALLBACK_API_KEY")
-	DSN := safeEnv("DSN")
-	REDIS_HOST := getEnv("REDIS_HOST", "cache")
-	REDIS_PORT := getEnv("REDIS_PORT", "6379")
-	INSTANCE_UUID = uuid.New().String()
-	debug, present := os.LookupEnv("DEBUG")
-
-	if !present {
-		DEBUG = false
-	} else {
-		DEBUG, _ = strconv.ParseBool(debug)
-	}
-
-	var err error
 	logger, err := zap.NewProduction()
 	if err != nil {
 		panic("Failed to get logger")
 	}
 	sugar := logger.Sugar()
+
+	cfg, errs := config.InitConfig()
+	if errs != nil {
+		for _, err := range errs {
+			sugar.Errorln(err)
+		}
+		panic("Failed to init config")
+	}
+	defer cfg.Shutdown()
 
 	e := echo.New()
 	e.Use(middleware.CORS())
@@ -80,7 +36,7 @@ func main() {
 				"request_id", "req_"+reqId,
 			)
 
-			cc := &Context{c, logger, reqId}
+			cc := &shared.Context{Context: c, Log: logger, Reqid: reqId, Cfg: cfg}
 			return next(cc)
 		}
 	})
@@ -94,255 +50,9 @@ func main() {
 			return c.String(500, "Internal Server Error")
 		},
 	}))
-	SQL_CLIENT, _ = sql.Open("mysql", DSN)
-	err = SQL_CLIENT.Ping()
-	if err != nil {
-		sugar.Error(err.Error())
-	}
-	defer SQL_CLIENT.Close()
-
-	REDIS_CLIENT = redis.NewClient(&redis.Options{
-		Addr:     fmt.Sprintf("%s:%s", REDIS_HOST, REDIS_PORT),
-		Password: "",
-		DB:       0,
-	})
-	defer REDIS_CLIENT.Close()
-
-	e.POST("/v1/chat/completions", func(c echo.Context) error {
-		cc := c.(*Context)
-		defer func() {
-			_ = cc.log.Sync()
-		}()
-		request, preprocessError := preprocessOpenaiRequest(cc, ENDPOINTS.CHAT)
-		if preprocessError != nil {
-			return cc.String(preprocessError.StatusCode, preprocessError.Error())
-		}
-
-		res, err := queryMiners(cc, request)
-		if err != nil {
-			cc.log.Warnw(
-				"Failed request, most likely un-recoverable. Not sending to fallback",
-				"error", err.Error(),
-				"final", "true",
-				"status", "failed",
-				"duration", fmt.Sprintf("%d", time.Since(request.StartTime)/time.Millisecond),
-			)
-			return c.JSON(500, OpenAIError{
-				Message: err.Error(),
-				Object:  "error",
-				Type:    "InternalServerError",
-				Code:    500,
-			})
-		}
-		go saveRequest(res, request, cc.log)
-		if res.Success {
-			return c.String(200, "")
-		}
-
-		if len(res.Data.Chat.Responses) > 15 {
-			cc.log.Warnw(
-				"failed request mid stream, canceling request",
-				"uid", res.Miner.Uid,
-				"hotkey", res.Miner.Hotkey,
-				"coldkey", res.Miner.Coldkey,
-				"error", res.Error,
-				"final", "true",
-				"status", "partial",
-				"duration", fmt.Sprintf("%d", time.Since(request.StartTime)/time.Millisecond),
-			)
-			return c.JSON(500, OpenAIError{
-				Message: "Failed mid-generation, please retry",
-				Object:  "error",
-				Type:    "InternalServerError",
-				Code:    500,
-			})
-		}
-		cc.log.Warnw(
-			"failed request, sending to fallback",
-			"uid", res.Miner.Uid,
-			"hotkey", res.Miner.Hotkey,
-			"coldkey", res.Miner.Coldkey,
-			"error", res.Error,
-		)
-		if DEBUG {
-			return c.String(500, "")
-		}
-		qerr := QueryFallback(cc, request)
-		if qerr != nil {
-			cc.log.Warnw(
-				"Failed fallback",
-				"error", qerr.Error(),
-				"final", "true",
-				"status", "failed",
-				"duration", fmt.Sprintf("%d", time.Since(request.StartTime)/time.Millisecond),
-			)
-			return c.JSON(503, OpenAIError{
-				Message: qerr.Error(),
-				Object:  "error",
-				Type:    "APITimeoutError",
-				Code:    qerr.StatusCode,
-			})
-		}
-
-		return c.String(200, "")
-	})
-
-	e.POST("/v1/completions", func(c echo.Context) error {
-		cc := c.(*Context)
-		defer func() {
-			_ = cc.log.Sync()
-		}()
-		request, preprocessError := preprocessOpenaiRequest(cc, ENDPOINTS.COMPLETION)
-		if preprocessError != nil {
-			return cc.String(preprocessError.StatusCode, preprocessError.Error())
-		}
-
-		res, err := queryMiners(cc, request)
-		if err != nil {
-			cc.log.Warnw(
-				"Failed request, most likely un-recoverable. Not sending to fallback",
-				"status", "failed",
-				"error", err.Error(),
-				"final", "true",
-			)
-			return c.JSON(500, OpenAIError{
-				Message: err.Error(),
-				Object:  "error",
-				Type:    "InternalServerError",
-				Code:    500,
-			})
-		}
-		go saveRequest(res, request, cc.log)
-		if res.Success {
-			return c.String(200, "")
-		}
-
-		if len(res.Data.Completion.Responses) > 15 {
-			cc.log.Warnw(
-				"failed request mid stream, canceling request",
-				"uid", res.Miner.Uid,
-				"hotkey", res.Miner.Hotkey,
-				"coldkey", res.Miner.Coldkey,
-				"error", res.Error,
-				"final", "true",
-				"status", "partial",
-				"duration", fmt.Sprintf("%d", time.Since(request.StartTime)/time.Millisecond),
-			)
-			return c.JSON(500, OpenAIError{
-				Message: "Failed mid-generation, please retry",
-				Object:  "error",
-				Type:    "InternalServerError",
-				Code:    500,
-			})
-		}
-
-		cc.log.Warnw(
-			"failed request, sending to fallback",
-			"uid", res.Miner.Uid,
-			"hotkey", res.Miner.Hotkey,
-			"coldkey", res.Miner.Coldkey,
-			"error", res.Error,
-		)
-		if DEBUG {
-			return c.String(500, "")
-		}
-		qerr := QueryFallback(cc, request)
-		if qerr != nil {
-			cc.log.Warnw("Failed fallback", "error", qerr.Error(), "final", "true", "status", "failed")
-			return c.JSON(503, OpenAIError{
-				Message: qerr.Error(),
-				Object:  "error",
-				Type:    "APITimeoutError",
-				Code:    qerr.StatusCode,
-			})
-		}
-
-		return c.String(200, "")
-	})
-
-	e.POST("/v1/images/generations", func(c echo.Context) error {
-		cc := c.(*Context)
-		defer func() {
-			_ = cc.log.Sync()
-		}()
-		request, preprocessError := preprocessOpenaiRequest(cc, ENDPOINTS.IMAGE)
-		if preprocessError != nil {
-			return cc.String(preprocessError.StatusCode, preprocessError.Error())
-		}
-		res, err := queryMiners(cc, request)
-
-		go saveRequest(res, request, cc.log)
-
-		if err != nil {
-			cc.log.Warnw("Failed request, sending to fallback server", "error", err.Error())
-			return c.String(500, err.Error())
-		}
-
-		if !res.Success {
-			cc.log.Warnf(
-				"Miner %d: %s %s\n Failed request\n",
-				res.Miner.Uid,
-				res.Miner.Hotkey,
-				res.Miner.Coldkey,
-			)
-			return c.String(
-				500,
-				fmt.Sprintf("Miner UID %d Failed Request. Try Again.", res.Miner.Uid),
-			)
-		}
-
-		// Send the image response - OpenAI image object
-		return c.JSON(200, map[string]interface{}{
-			"b64_json": res.Data.Image,
-		})
-	})
-
-	e.GET("/v1/models", func(c echo.Context) error {
-		cc := c.(*Context)
-		defer cc.log.Sync()
-
-		rows, err := SQL_CLIENT.Query(`
-			SELECT name, DATE_FORMAT(created_at, '%Y-%m-%d %H:%i:%s') as created_at 
-			FROM model WHERE enabled = 1
-		`)
-		if err != nil {
-			cc.log.Errorw("Failed to get models", "error", err.Error())
-			return c.String(500, "Failed to get models")
-		}
-		defer rows.Close()
-
-		var models []Model
-		for rows.Next() {
-			var model Model
-			var createdAtStr string
-			if err := rows.Scan(&model.ID, &createdAtStr); err != nil {
-				cc.log.Errorw("Failed to scan model row", "error", err.Error())
-				return c.String(500, "Failed to get models")
-			}
-
-			createdAt, err := time.Parse("2006-01-02 15:04:05", createdAtStr)
-			if err != nil {
-				cc.log.Errorw("Failed to parse created_at", "error", err.Error())
-				return c.String(500, "Failed to get models")
-			}
-
-			model.Object = "model"
-			model.Created = createdAt.Unix()
-			model.OwnedBy = strings.Split(model.ID, "/")[0]
-
-			models = append(models, model)
-		}
-
-		if err = rows.Err(); err != nil {
-			cc.log.Error("Error iterating model rows", "error", err.Error())
-			return c.String(500, "Failed to get models")
-		}
-
-		return c.JSON(200, ModelList{
-			Object: "list",
-			Data:   models,
-		})
-	})
+	e.POST("/v1/chat/completions", routes.ChatRequest)
+	e.POST("/v1/completions", routes.CompletionRequest)
+	e.GET("/v1/models", routes.Models)
 
 	e.Logger.Fatal(e.Start(":80"))
 }
