+++ docker-compose.yml
@@ -40,7 +40,7 @@ services:
       context: ./miner-cache
       platforms: 
       - linux/amd64
-      - linux/arm64
+      #- linux/arm64
     volumes:
       - ./blocked_keys.txt:/app/blocked_keys.txt
     restart: always
